#154 FIX: Altera modal de registro de usuário
